refactor: :recycle: move statNames to lib.js
